requirements.txt
aiosqlite==0.20.0
beautifulsoup4==4.12.3
soupsieve==2.6
lxml==5.3.0
aiogram==3.12.0
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2